Added the contact 3d + fixed two bugs in 18.04

See merge request loco-3d/crocoddyl!194

+++ include/crocoddyl/core/optctrl/shooting.hpp
@@ -25,7 +25,7 @@ class ShootingProblem {
   void rollout(const std::vector<Eigen::VectorXd>& us, std::vector<Eigen::VectorXd>& xs);
   std::vector<Eigen::VectorXd> rollout_us(const std::vector<Eigen::VectorXd>& us);
 
-  long unsigned int get_T() const;
+  unsigned int get_T() const;
   const Eigen::VectorXd& get_x0() const;
 
   std::vector<ActionModelAbstract*>& get_runningModels();
@@ -40,7 +40,7 @@ class ShootingProblem {
 
  protected:
   void allocateData();
-  long unsigned int T_;
+  unsigned int T_;
   Eigen::VectorXd x0_;
 
  private:
